pyproject.toml
[project]
name = "projectairqualitytraffic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair==4.2.2",
    "apache-airflow-providers-postgres>=6.6.1",
    "boto3>=1.42.77",
    "dotenv>=0.9.9",
    "folium>=0.20.0",
    "lxml>=6.0.2",
    "numpy>=2.4.2",
    "openmeteo-requests>=1.7.5",
    "pandas>=3.0.1",
    "polars>=1.38.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pyproj>=3.7.2",
    "requests-cache>=1.3.1",
    "retry-requests>=2.0.0",
    "s3fs>=0.4.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "statsmodels>=0.14.6",
    "streamlit>=1.19.0",
    "xgboost>=3.2.0",
]
